requirements.txt
Django>=4.2,<5.0
Pillow>=10.0.0
psycopg2-binary>=2.9.0
python-decouple>=3.8
whitenoise>=6.5.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
django-ckeditor>=6.7.0
django-storages>=1.14.0
boto3>=1.34.0
django-modeltranslation>=0.18.12

